+++ Assets/_Scripts/ScriptableObjects/TilePaletteScriptableObject.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Tilemaps;
@@ -5,8 +6,9 @@ using UnityEngine.Tilemaps;
 namespace _Scripts.ScriptableObjects
 {
     [CreateAssetMenu]
+    [Serializable]
     public class TilePaletteScriptableObject : ScriptableObject
     {
-        public List<Tile> tilePalette;
+        public List<Tile> tiles;
     }
 }