+++ Assets/_Scripts/_GradientNoise/CellMapGenerator.cs
@@ -71,21 +71,21 @@ namespace _Scripts._GradientNoise
             {
                 // all indoor cells are cave
                 cell.Biom = Biom.Cave;
-
+            
                 // Can be generated with Open Simplex Noise, Perlin Noise
                 var val = ValueGenerator.Evaluate(cell.CellIndex.x, cell.CellIndex.y, mountainLayerSettings);
-
+            
                 cell.Asset = val == 1
                     ? new CellAsset(CellAsset.AssetType.MassiveRock)
                     : new CellAsset(CellAsset.AssetType.Cavity);
             }
-
+            
             // open terrain layer generation
             foreach (var cell in _outdoorCells)
             {
                 // Can be generated with Open Simplex Noise, Perlin Noise
                 var val = ValueGenerator.Evaluate(cell.CellIndex.x, cell.CellIndex.y, outdoorBiomSetting);
-
+            
                 if (val == 1)
                 {
                     // cell is Meadows
@@ -95,12 +95,12 @@ namespace _Scripts._GradientNoise
                 {
                     // cell is Woods
                     cell.Biom = Biom.Woods;
-
+            
                     var value = prng.Next(101);
                     var trees = assetGenerationSettings.treePercentage;
                     var bushes = assetGenerationSettings.bushPercentage;
                     var gras = assetGenerationSettings.grassPercentage;
-
+            
                     if (trees + bushes + gras > 100)
                     {
                         Debug.LogError("More than 100% Assets.");
@@ -118,7 +118,7 @@ namespace _Scripts._GradientNoise
                     }
                 }
             }
-
+            
             // Get the values of the neighbours
             // If one or more neighbours are different from the current cell, make the current cell a wall
             foreach (var cell in _indoorCells)
@@ -130,22 +130,22 @@ namespace _Scripts._GradientNoise
                     {
                         int xPos = cell.CellIndex.x + x;
                         int yPos = cell.CellIndex.y + y;
-
+            
                         // skip the incoming cell, and cell coordinates that are not in the map
                         if ((xPos == cell.CellIndex.x && yPos == cell.CellIndex.y) || xPos < 0 || yPos < 0 ||
                             xPos >= resolution.x || yPos >= resolution.y)
                         {
                             continue;
                         }
-
+            
                         var neighbourVal = _cellMap[xPos, yPos].Indoors;
                         var neighbourAsset = _cellMap[xPos, yPos].Asset;
-
+            
                         if (neighbourVal == false)
                         {
                             cell.Asset = new CellAsset(CellAsset.AssetType.Wall);
                         }
-
+            
                         if (cell.Asset.Type == CellAsset.AssetType.MassiveRock &&
                             neighbourAsset.Type == CellAsset.AssetType.Cavity)
                         {